move all ffi into a mod
