+++ engine_store_ffi/src/core/common.rs
@@ -37,7 +37,10 @@ pub use yatp::{
 };
 
 pub(crate) use crate::{
-    gen_engine_store_server_helper, interfaces::root::DB::EngineStoreApplyRes, name_to_cf,
-    ColumnFamilyType, EngineStoreServerHelper, RaftCmdHeader, RawCppPtr, TiFlashEngine,
-    WriteCmdType, WriteCmds,
+    ffi::{
+        gen_engine_store_server_helper, interfaces::root::DB::EngineStoreApplyRes, name_to_cf,
+        ColumnFamilyType, EngineStoreServerHelper, RaftCmdHeader, RawCppPtr, WriteCmdType,
+        WriteCmds,
+    },
+    TiFlashEngine,
 };
