+++ engine_store_ffi/src/ffi/encryption_impls.rs
@@ -2,7 +2,7 @@
 
 use engine_traits::{EncryptionKeyManager, EncryptionMethod, FileEncryptionInfo};
 
-use crate::{
+use super::{
     get_engine_store_server_helper,
     interfaces::root::DB::{
         BaseBuffView, EncryptionMethod as EncryptionMethodImpl, FileEncryptionInfoRaw,
