+++ new-mock-engine-store/src/mock_page_storage.rs
@@ -6,7 +6,7 @@ use std::{
     sync::{atomic::AtomicU64, Arc, RwLock},
 };
 
-pub use engine_store_ffi::{
+pub use engine_store_ffi::ffi::{
     interfaces::root::DB as ffi_interfaces, BaseBuffView, CppStrWithView, EngineStoreServerHelper,
     PageAndCppStrWithView, RaftStoreProxyFFIHelper, RawCppPtr, RawCppPtrCarr, RawVoidPtr,
 };
